requirements.txt
streamlit
torch
torchvision
opencv-python
Pillow
numpy
matplotlib
scikit-learn
datasets
colorama
streamlit-webrtc
